+++ changelog.md
@@ -1,6 +1,7 @@
 # v0.1-beta.2
 - Added "Refresh Interval" option in settings
 - Improved the mod's settings
+- Rebranded "Tags" as "Names"
 
 # v0.1-beta.1
 - Rewrote ALL of the notification logic